requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
black==25.12.0
cfgv==3.5.0
click==8.3.1
distlib==0.4.0
fastapi==0.124.2
filelock==3.20.0
flake8==7.3.0
greenlet==3.3.0
h11==0.16.0
identify==2.6.15
idna==3.11
jose==1.0.0
logging==0.4.9.6
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.1
psycopg2==2.9.11
pycodestyle==2.14.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
PyYAML==6.0.3
ruff==0.14.8
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
virtualenv==20.35.4
